requirements.txt
# EmoTranslator Dependencies
# Core dependencies for the Emoji Mood Translator application

# Web framework for the UI
streamlit>=1.28.0

# OpenAI API client for GPT integration
openai>=1.3.0

# Environment variable management
python-dotenv>=1.0.0

# Clipboard operations
pyperclip>=1.8.2

# Development dependencies (uncomment for development)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Production dependencies (uncomment for deployment)
# gunicorn>=21.2.0